requirements.txt
streamlit
transformers
sentencepiece
torch
PyMuPDF
newspaper3k
